moved grouping self loops related code into a separate file
